+++ lib/shrine/plugins/activerecord.rb
@@ -26,7 +26,7 @@ class Shrine
     # you should first disable these transactions for those tests.
     #
     # If you want to put some parts of this lifecycle into a background job, see
-    # the background_helpers plugin.
+    # the backgrounding plugin.
     #
     # Additionally, any Shrine validation errors will added to ActiveRecord's
     # errors upon validation. Note that if you want to validate presence of the
@@ -64,7 +64,7 @@ class Shrine
       end
 
       module AttacherClassMethods
-        # Needed by the background_helpers plugin.
+        # Needed by the backgrounding plugin.
         def find_record(record_class, record_id)
           record_class.find(record_id)
         end